accelStepper project working!
accelStepper is working neatly!. This is my own implementation of
the accelStepper library, with some things missing and some
additional things. Still a couple of funcions missing: change
max speed and acceleration.

+++ accelStepper/timers.h
@@ -13,6 +13,6 @@ void speed_timer_set_raw(uint16_t c);
 
 void aux_timer_init(void);
 
-void aux_timer_set(uint8_t state, uint16_t t);
+void aux_timer_set(uint8_t state, uint8_t t);
 
 #endif /* TIMERS_H */
